@@ -9,13 +9,14 @@
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"lint": "prettier --plugin-search-dir . --check . && eslint .",
-		"format": "prettier --plugin-search-dir . --write ."
+		"format": "prettier --plugin-search-dir . --write .",
+		"prisma": "npx prisma migrate dev --name init"
 	},
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^2.0.0",
 		"@sveltejs/adapter-node": "^1.2.3",
 		"@sveltejs/adapter-vercel": "^3.0.1",
-		"@sveltejs/kit": "^1.5.0",
+		"@sveltejs/kit": "^1.20.4",
 		"@typescript-eslint/eslint-plugin": "^5.45.0",
 		"@typescript-eslint/parser": "^5.45.0",
 		"autoprefixer": "^10.4.14",
@@ -24,13 +25,15 @@
 		"eslint-plugin-svelte3": "^4.0.0",
 		"postcss": "^8.4.23",
 		"prettier": "^2.8.0",
-		"prettier-plugin-svelte": "^2.8.1",
-		"svelte": "^3.54.0",
-		"svelte-check": "^3.0.1",
+		"prettier-plugin-svelte": "^2.10.1",
+		"prettier-plugin-tailwindcss": "^0.3.0",
+		"prisma": "^4.16.2",
+		"svelte": "^4.0.0",
+		"svelte-check": "^3.4.3",
 		"tailwindcss": "^3.3.1",
 		"tslib": "^2.4.1",
 		"typescript": "^5.0.0",
-		"vite": "^4.2.0"
+		"vite": "4.3.9"
 	},
 	"type": "module",
 	"dependencies": {
@@ -40,6 +43,17 @@
 		"@fortawesome/free-brands-svg-icons": "^6.4.0",
 		"@fortawesome/free-regular-svg-icons": "^6.4.0",
 		"@fortawesome/free-solid-svg-icons": "^6.4.0",
-		"svelte-awesome": "^3.2.0"
+		"@prisma/client": "^4.16.2",
+		"@types/dompurify": "^3.0.2",
+		"@types/marked": "^5.0.0",
+		"@types/node": "^20.3.3",
+		"buffer": "^6.0.3",
+		"dompurify": "^3.0.4",
+		"marked": "^5.1.0",
+		"minio": "^7.1.1",
+		"parse-md": "^3.0.3",
+		"stream": "^0.0.2",
+		"svelte-awesome": "^3.2.0",
+		"util": "^0.12.5"
 	}
 }
