@@ -1,3 +1,3 @@
-<div class='w-full my-12 flex items-center justify-center'>
-	<div class='w-1/4 h-[2px] rounded-[50px] bg-grayedOut_dark'></div>
+<div class="my-12 flex w-full items-center justify-center">
+	<div class="h-[2px] w-1/4 rounded-[50px] bg-grayedOut_dark" />
 </div>
