@@ -1,21 +1,27 @@
-<script lang='ts'>
+<script lang="ts">
 	export let text: string;
 	export let img: string;
 	export let imgAlt: string;
 	export let mode: 'left' | 'right' = 'left';
 </script>
 
-<div class='flex items-center lg:justify-between gap-4 flex-col lg:flex-row px-8 md:px-16 lg:px-24 w-full'>
+<div
+	class="flex w-full flex-col items-center gap-4 px-8 md:px-16 lg:flex-row lg:justify-between lg:px-24"
+>
 	{#if mode === 'left'}
-		<p class='text-center w-full lg:text-right lg:w-1/2'>
+		<p class="w-full text-center lg:w-1/2 lg:text-right">
 			{text}
 		</p>
 	{/if}
 
-	<img src={img} alt={imgAlt} class='w-full md:w-3/4 lg:w-2/5 rounded-[30px] border-grayedOut_dark dshadow border-[5px]'>
+	<img
+		src={img}
+		alt={imgAlt}
+		class="dshadow w-full rounded-[30px] border-[5px] border-grayedOut_dark md:w-3/4 lg:w-2/5"
+	/>
 
 	{#if mode === 'right'}
-		<p class='text-center w-full lg:text-left lg:w-1/2'>
+		<p class="w-full text-center lg:w-1/2 lg:text-left">
 			{text}
 		</p>
 	{/if}
