@@ -2,12 +2,19 @@
 	import FooterSocialIcons from './FooterSocialIcons.svelte';
 </script>
 
-<div class='flex items-center justify-center gap-4'>
-	<img src='/img/logo/netzwerk.png' alt='Cubyx Network Logo' class='h-[7rem] md:h-[10rem] lg:h-[12.5rem]'>
-	<div class='flex flex-col gap-2'>
+<div class="flex items-center justify-center gap-4">
+	<img
+		src="/img/logo/netzwerk.png"
+		alt="Cubyx Network Logo"
+		class="h-[7rem] md:h-[10rem] lg:h-[12.5rem]"
+	/>
+	<div class="flex flex-col gap-2">
 		<div>
-			<h3 class='leading-tight'>Cubyx Network</h3>
-			<h4 class='text-grayedOut_dark leading-tight'>Quadratisch<span class='text-accent'>.</span> Praktisch<span class='text-accent'>.</span> Gut<span class='text-accent'>.</span></h4>
+			<h3 class="leading-tight">Cubyx Network</h3>
+			<h4 class="leading-tight text-grayedOut_dark">
+				Quadratisch<span class="text-accent">.</span> Praktisch<span class="text-accent">.</span>
+				Gut<span class="text-accent">.</span>
+			</h4>
 		</div>
 		<FooterSocialIcons />
 	</div>
