@@ -10,12 +10,12 @@
 	import { faSpotify } from '@fortawesome/free-brands-svg-icons/faSpotify';
 </script>
 
-<div class='flex items-center gap-3 md:gap-4'>
-	<FooterSocialIcon icon={faDiscord} href='https://cubyx.eu/discord' />
-	<FooterSocialIcon icon={faTwitch} href='https://cubyx.eu/twitch' />
-	<FooterSocialIcon icon={faYoutube} href='https://cubyx.eu/youtube' />
-	<FooterSocialIcon icon={faSpotify} href='https://cubyx.eu/podcast' />
-	<FooterSocialIcon icon={faMastodon} href='https://cubyx.eu/mastodon' />
-	<FooterSocialIcon icon={faReddit} href='https://cubyx.eu/reddit' />
-	<FooterSocialIcon icon={faEarthEurope} href='https://cubyx.eu/planetmc' />
+<div class="flex items-center gap-3 md:gap-4">
+	<FooterSocialIcon icon={faDiscord} href="https://cubyx.eu/discord" />
+	<FooterSocialIcon icon={faTwitch} href="https://cubyx.eu/twitch" />
+	<FooterSocialIcon icon={faYoutube} href="https://cubyx.eu/youtube" />
+	<FooterSocialIcon icon={faSpotify} href="https://cubyx.eu/podcast" />
+	<FooterSocialIcon icon={faMastodon} href="https://cubyx.eu/mastodon" />
+	<FooterSocialIcon icon={faReddit} href="https://cubyx.eu/reddit" />
+	<FooterSocialIcon icon={faEarthEurope} href="https://cubyx.eu/planetmc" />
 </div>
