@@ -1,1 +1,1 @@
-<div class='h-24'></div>
+<div class="h-24" />
