@@ -1,7 +1,10 @@
 <script lang="ts">
-    export let justify: "start" | "end";
+	export let justify: 'start' | 'end';
 </script>
 
-<div class="sticky flex items-center justify-start {justify === 'end' && 'justify-end'} gap-5 w-1/3 hidden lg:inline-flex">
-    <slot></slot>
+<div
+	class="sticky flex items-center justify-start {justify === 'end' &&
+		'justify-end'} hidden w-1/3 gap-5 lg:inline-flex"
+>
+	<slot />
 </div>
